+++ src/main/kotlin/xyz/xenondevs/nova/tileentity/impl/energy/PowerCell.kt
@@ -19,7 +19,7 @@ import java.util.*
 
 open class PowerCell(
     creative: Boolean,
-    val maxEnergy: Int,
+    val maxEnergy: Long,
     uuid: UUID,
     data: CompoundElement,
     material: NovaMaterial,
@@ -57,10 +57,10 @@ open class PowerCell(
     
 }
 
-private val BASIC_CAPACITY = NovaConfig[NovaMaterialRegistry.BASIC_POWER_CELL].getInt("capacity")!!
-private val ADVANCED_CAPACITY = NovaConfig[NovaMaterialRegistry.ADVANCED_POWER_CELL].getInt("capacity")!!
-private val ELITE_CAPACITY = NovaConfig[NovaMaterialRegistry.ELITE_POWER_CELL].getInt("capacity")!!
-private val ULTIMATE_CAPACITY = NovaConfig[NovaMaterialRegistry.ULTIMATE_POWER_CELL].getInt("capacity")!!
+private val BASIC_CAPACITY = NovaConfig[NovaMaterialRegistry.BASIC_POWER_CELL].getLong("capacity")!!
+private val ADVANCED_CAPACITY = NovaConfig[NovaMaterialRegistry.ADVANCED_POWER_CELL].getLong("capacity")!!
+private val ELITE_CAPACITY = NovaConfig[NovaMaterialRegistry.ELITE_POWER_CELL].getLong("capacity")!!
+private val ULTIMATE_CAPACITY = NovaConfig[NovaMaterialRegistry.ULTIMATE_POWER_CELL].getLong("capacity")!!
 
 class BasicPowerCell(
     uuid: UUID,
@@ -134,7 +134,7 @@ class CreativePowerCell(
     armorStand: FakeArmorStand,
 ) : PowerCell(
     true,
-    Int.MAX_VALUE,
+    Long.MAX_VALUE,
     uuid,
     data,
     material,
